refactor(Utils): update file paths and module imports
- Update file paths and module imports across multiple files
- Move constants and configuration files to Utils directory- Adjust import statements to use new file paths
- Update package.json scripts
