+++ Src/Slash-Commands/contributors.js
@@ -1,7 +1,7 @@
 const {SlashCommandBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle} = require('discord.js');
 const axios = require('axios');
-const {CONTRIBUTORS_URL} = require('../../utils/config-global.js');
-const {MESSAGE_COLLECTOR_TIMEOUT, remainingTime, createEmbed, EMBED_COLORS} = require('../../utils/constants.js');
+const {CONTRIBUTORS_URL} = require('../../Utils/bot/config-global');
+const {MESSAGE_COLLECTOR_TIMEOUT, remainingTime, createEmbed, EMBED_COLORS} = require('../../Utils/cmds/constants.js');
 
 module.exports = {
     data: new SlashCommandBuilder()
