+++ Utils/bot/constants.js
@@ -1,0 +1,13 @@
+/**
+ * NOTICE:
+ * This file contains configuration constants used throughout the application.
+ * Any changes to this file should be carefully reviewed to ensure they do not
+ * affect other parts of the application that rely on these configuration values.
+ */
+
+// Activity rotation interval in milliseconds (default: 30 seconds)
+const ACTIVITY_ROTATION_INTERVAL = 30000;
+
+module.exports = {
+    ACTIVITY_ROTATION_INTERVAL
+};
